+++ decomp/ST.exe/functions/006BD4B0/decomp.c
@@ -10,8 +10,8 @@ void FUN_006bd4b0(void)
      ((undefined8 *)DAT_00854f44[0x30] != (undefined8 *)0x0)) {
     FUN_006cea60((undefined8 *)DAT_00854f44[0x30],DAT_00854f44[3],
                  (undefined8 *)
-                 (DAT_00854f44[2] * DAT_00854fe0 + DAT_00854f44[1] + (int)DAT_00854f00),DAT_00854fe0
-                 ,DAT_00854f44[3],DAT_00854f44[4]);
+                 (DAT_00854f44[2] * (int)DAT_00854fe0 + DAT_00854f44[1] + (int)DAT_00854f00),
+                 (int)DAT_00854fe0,DAT_00854f44[3],DAT_00854f44[4]);
   }
   DAT_00854edc = DAT_00854f44[0x1a];
   if ((int)DAT_00854edc < 5) {
